Fix and move examples in package to ease the use after pypi install
